Enable caching of (non-paginated) single- and multi-element queries
Not sure how many disclaimers to offer, here! I’d love someone to
review this, but the original authors are undoubtedly in a better
position to execute a robust, sensitive solution.

In any case, this seems to substantially cut down on the time for
subsequent fetches of data over AJAX, and will (in the least) offer
some relief for applications that may experience either heavy load or
attempted abuse. I can imagine that the performance boost will be more
significant for larger queries, but for fairly basic queries, these
improvements reduced almost all responses to around 100ms.