Update README.md
had some trouble using the setup example on a freshly installed craft due to the lack of ->getParsedContent() method being applied to parse the body content, which by default in the provided stock news section is a rich text field, and without the getParsedContent you only get a empty object. Also removed the $entry.summary, since a default news does not seem to have one. I understand that this is only an example, but it would be great to just plug the code and see it working on a freshly installed craft